% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{j_encode}
\alias{j_encode}
\title{Encoding Function}
\usage{
j_encode(data, json = TRUE)
}
\arguments{
\item{data}{A list or data.frame to encode.}

\item{json}{Defaults to TRUE. If TRUE, the output will be a base64 encoded JSON, else, the output will be an encoded string.}
}
\description{
Used to encode a data.frame or list. By default, the output will be a base64 encoded JSON.
}
\examples{
# Transform a data.frame to an encoded JSON string
df <- iris
encoded <- j_encode(df, json = TRUE)

}
\keyword{encode}
