% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_file.R
\name{j_save}
\alias{j_save}
\title{Save encoded string}
\usage{
j_save(str, file = NULL)
}
\arguments{
\item{str}{An object containing the string to save.}

\item{file}{Optional. The name and format of the file. If no value is provided, the function will use the name of the object passed as str argument and save it as a txt file <object.txt> on your working directory.}
}
\description{
Mainly used to save an encoded output in a file on your working directory.
}
\details{
This function acts as a wrapper for write.table function, intended to save the output of the
j_encode function in a text file (useful when the result is a long string).

The wrapper strips the row and col names and also removes the quotes at the start and end of the string.
}
\keyword{save}
