% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{j_decode}
\alias{j_decode}
\title{Decoding Function}
\usage{
j_decode(str, json = TRUE)
}
\arguments{
\item{str}{The string to be decoded.}

\item{json}{Defaults to TRUE. If TRUE, the function expects str to be an encoded json and will return a data.frame or list, depending on JSON structure. If FALSE, the function will return an string.}
}
\description{
Used to decode a base64 string. By default the function expects an encoded json.
}
\examples{
# Decode an encoded string:

str <- "SGVsbG8gV29ybGQh"
j_decode(str, json = FALSE)

# Decode an encoded json:

encoded_json <- "W3sibXNnIjogIkhlbGxvIFdvcmxkISIsICJqc29uIjogdHJ1ZX1d"
j_decode(encoded_json)
}
\keyword{decode}
