% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{j_save}
\alias{j_save}
\title{Save file Function}
\usage{
j_save(data, filename)
}
\arguments{
\item{data}{The data to be saved}

\item{filename}{A character string representing the name of the file.}
}
\description{
Used to save data to a file in the working directory using UTF-8 encoding.
}
\examples{
# Write a JSON string to a UTF-8 encoded file

data <- "[{msg: 'Hello R users!', type: 'example'}]"
j_save(data, 'data.json')
}
\keyword{file}
\keyword{write}
